+++ crates/task-impls/src/harness.rs
@@ -31,109 +31,6 @@ impl<TYPES: NodeType> TaskState for TestHarnessState<TYPES> {
     }
 }
 
-// A much more tightly-choreographed version of `TestHarnessState`.
-//
-// A `TestScript` is a sequence of pairs (input sequence, output sequence).
-type TestScript<TYPES, S> = Vec<TestScriptStage<TYPES, S>>;
-
-pub struct TestScriptStage<TYPES: NodeType, S: TaskState<Event = HotShotEvent<TYPES>>> {
-    pub inputs: Vec<HotShotEvent<TYPES>>,
-    pub outputs: Vec<Predicate<HotShotEvent<TYPES>>>,
-    pub asserts: Vec<Predicate<S>>,
-}
-
-pub struct Predicate<INPUT> {
-    pub function: Box<dyn for<'a> Fn(&'a INPUT) -> bool>,
-    pub info: String,
-}
-
-impl<INPUT> std::fmt::Debug for Predicate<INPUT> {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> Result<(), std::fmt::Error> {
-        write!(f, "{}", self.info)
-    }
-}
-
-/// `run_test_script` reads a pair (inputs, outputs) in the vector,
-/// broadcasts all given inputs (in order) and waits to receive all outputs (in order).
-/// It moves on to the next pair only *after* it has finished receiving all expected outputs.
-pub async fn run_test_script<TYPES, S: TaskState<Event = HotShotEvent<TYPES>>>(
-    mut script: TestScript<TYPES, S>,
-    state: S,
-) where
-    TYPES: NodeType,
-    S: Send + 'static,
-{
-    let registry = Arc::new(TaskRegistry::default());
-
-    let (test_input, task_receiver) = broadcast(1024);
-    // let (task_input, mut test_receiver) = broadcast(1024);
-
-    let task_input = test_input.clone();
-    let mut test_receiver = task_receiver.clone();
-
-    let mut task = Task::new(
-        task_input.clone(),
-        task_receiver.clone(),
-        registry.clone(),
-        state,
-    );
-
-    for (stage_number, stage) in script.iter_mut().enumerate() {
-        tracing::debug!("Beginning test stage {}", stage_number);
-        for input in &mut *stage.inputs {
-            if !task.state_mut().filter(&input) {
-                tracing::debug!("Test sent: {:?}", input);
-
-                if let Some(res) = S::handle_event(input.clone(), &mut task).await {
-                    task.state_mut().handle_result(&res).await;
-                }
-            }
-        }
-
-        for expected in &stage.outputs {
-            let timeout_error = format!(
-                "Timeout while waiting for output at stage {}: {:?}",
-                stage_number, expected
-            );
-
-            if let Ok(received_output) =
-                async_timeout(Duration::from_secs(2), test_receiver.recv_direct())
-                    .await
-                    .expect(&timeout_error)
-            {
-                let predicate = &expected.function;
-                tracing::debug!("Test received: {:?}", received_output);
-                assert!(
-                    predicate(&received_output),
-                    "Output failed to satisfy {:?}",
-                    expected
-                );
-            } else {
-                panic!("{}", timeout_error);
-            }
-        }
-
-        for assert in &stage.asserts {
-            assert!(
-                task.test_state_with(&assert.function),
-                "TaskState assertion failed at stage {}: {:?}",
-                stage_number,
-                assert
-            );
-        }
-
-        // Once we've seen all expected outputs, we wait to see
-        // if there are any trailing unexpected outputs.
-        if let Ok(output) = async_timeout(Duration::from_secs(2), test_receiver.recv_direct()).await
-        {
-            panic!(
-                "Received unexpected output at stage {}: {:?}",
-                stage_number, output
-            );
-        }
-    }
-}
-
 /// Runs a test by building the task using `build_fn` and then passing it the `input` events
 /// and testing the make sure all of the `expected_output` events are seen
 ///
