+++ crates/task/src/task.rs
@@ -166,9 +166,9 @@ impl<S: TaskState + Send + 'static> Task<S> {
     pub fn state_mut(&mut self) -> &mut S {
         &mut self.state
     }
-
-    pub fn test_state_with(&self, predicate: &Box<dyn Fn(&S) -> bool>) -> bool {
-      predicate(&self.state)
+    /// Get an immutable reference to this tasks state
+    pub fn state(&self) -> &S {
+        &self.state
     }
 
     /// Spawn a new task adn register it.  It will get all events not seend
